// src/types.d.ts
interface ForestBotWebsocketClientOptions {
    mc_server: string
    websocket_url: string
    apiKey: string
}

interface MinecraftChatMessage {
    name: string
    message: string
    date: { String: string, Valid: boolean }
    mc_server: string
    uuid: string
}

interface DiscordChatMessage {
    message: string
    username: string
    timestamp: string
    server: string
    channel_id: string
    guild_id: string
    guild_name: string
}

interface MinecraftAdvancementMessage {
    username: string
    advancement: string
    time: number,
    mc_server: string,
    uuid: string
    id?: number | null
}

interface MinecraftPlayerJoinMessage {
    username: string
    uuid: string
    timestamp: string
    server: string
}

interface MinecraftPlayerLeaveMessage {
    username: string
    uuid: string
    timestamp: string
    server: string
}

interface MinecraftPlayerKillMessage {
    username: string
    uuid: string
    timestamp: string
    server: string
}

interface MinecraftPlayerDeathMessage {
    victim: string
    death_message: string

    //the object will be present only on pvp deaths and when recieving from websocket.
    //we do not use the object when sending a death to the api.
    //we send the murderer as a string, (username)
    murderer?: { String: string, Valid: boolean }
    time: number,
    type: "pve" | "pvp",
    mc_server: string,
    id?: number | null | undefined
    victimUUID: string,
    murdererUUID?: string
}

type messageActionTypes = "error" | "new_user" | "key-accepted" | "new_name" | "id" | "minecraft_chat" | "discord_chat" | "send_update_player_list" | "minecraft_advancement" | "minecraft_player_join" | "minecraft_player_leave" | "minecraft_player_kill" | "minecraft_player_death";
type inboundmessageDataTypes = DiscordChatMessage | MinecraftAdvancementMessage | MinecraftChatMessage | MinecraftPlayerDeathMessage | MinecraftPlayerKillMessage | MinecraftPlayerJoinMessage | MinecraftPlayerLeaveMessage | any;

// type outboundMessageActionTypes = 
// type outboundMessageDataTypes = any;

interface InBoundWebsocketMessage {
    client_id: string,
    action: messageActionTypes
    data: inboundmessageDataTypes
}

interface OutboundWebsocketMessage {
    client_id: string,
    action: messageActionTypes,
    data: inboundmessageDataTypes
}
// ...
interface Player {
    username: string,
    uuid: string,
    latency: number,
    server: string
}